feat: Update frontend to handle new registry types (#566)
This PR updates the frontend to work with the updated registry contract
types/methods. There are two parts to this change:
- Consuming the new types, mainly `rule` and `startBlock`, and writing
these types back during registration
- Adding the new `StartBlock::Continue` option, which allows developers
to resume the existing block stream.

I've attached images of the updated UI for reference. Note that updating
the contract filter is disabled with the new "Continue" option.


![image](https://github.com/near/queryapi/assets/11974624/d4069cc8-d7bf-4c66-9d33-ec9d51bfcc83)

![image](https://github.com/near/queryapi/assets/11974624/0b3ef7e4-6542-4b5a-86ad-14c8028565ed)