feat: Cache Database Credentials (#585)
Currently, Runner fetches DB credentials every invocation of runFunction
because provisioning of the user's DB occurs in Runner. As such, we
couldn't be sure if we could fetch credentials for the user's DB from
Hasura or not. The fetch is slow and increases the latency of any
indexers which leverage context.db.

I updated the code to cache the credentials locally after the first
successful fetch. This allows all future requests to succeed as these
credentials will never change after they are set. The changes were aimed
to be minimal as we intend to migrate provisioning out of Runner. In
fact, the changes aid in the effort as the credentials fetch is done
through the Provisioner class rather than separately in DmlHandler.