feat: Reduce Runner Logs and DB Writes (#589)
Runner logs excessively, both to the machine, and to postgres. Most of
these logs are not necessary or contain too much information.

The plan is to reduce logs the following ways.

1. Only output failure log for first failure (Instead of every 10
seconds)
2. Stop outputting "Running function on X block" logs to GCP
3. Stop outputting "Success: Ran function on X block" logs to GCP
4. Remove DB logs for hasura calls
5. Remove context.db logging

To support 1, Runner needs a way to know when a failure was the first.
So, it needs to remember the state of the Runner locally. This is a good
opportunity to introduce a refactor to stop setting of Runner status
every loop. We can instead only write when the status changes.

In addition, there is some overlap with this epic. So, to orient these
changes towards our intentions there, I will also introduce a
rudimentary log level into Runner. This value will be INFO by default,
and the extra logs will be at the DEBUG level. This ensures these logs
are automatically skipped while also adding some backbone for expected
future implementation.