fix: Cap Block Streamer Caching and Merge Redis Impl (#599)
Redis memory usage is pegged at 100% due to Block Streamer backfilling *
contract filter indexers.

In addition, Redis implementation of Coordinator and Block Streamer have
diverged. Coordinator is also not able to make successful calls to
Redis. While not a targeted fix, the hope is that merging the
implementation fixes the issue as Block Streamer is able to make calls
to Redis.

Before release, some manual intervention is necessary in dev and prod to
update versions of indexers to the current, to prevent unwanted behavior
from coordinator. I will do dev first, release the change, and then do
prod before a prod release.