fix: Substitution 'node-sql-parser' with a forked version until April 1st (Next Release) (#597)
A temporary change in our codebase. Replacing the usage of
node-sql-parser with kevin-node-sql-parser until April 1st. The reason
for this substitution is that the official release of node-sql-parser
lacks a version release for the additional SQL statements required for
our current project needs. In the interim, this forked version addresses
these shortcomings and allows us to incorporate the required SQL
statements. Please note that this is a temporary measure, and we plan to
revert to the official node-sql-parser version after April 1st, once the
required features are officially available.

See last comment for details