fix: Reduce requests made to Near Lake S3 (#665)
Each `BlockStream` uses its own dedicated `near-lake-framework`
instance, and hence manages its own connection with S3. This leads to
many duplicate S3 requests, particularly across the large majority of
Indexers which follow the network tip, which request the same block data
at the same time.

This PR introduces a shared S3 client to be used across all
`near-lake-framework` instances. `SharedLakeS3Client` ensures that
duplicate requests made within a short time-frame, including those made
in parallel, result in only a single request to S3.

## Cache Strategy

This implementation will mostly impact `BlockStream`s following the
network tip, i.e. `From Latest`. These streams will wait for new data in
Near Lake S3, and request it as soon as it is available, at the same
time. Therefore, it would be enough to cache the result alone, by the
time we actually prime the cache, all other requests would have missed
it and fired a request of their own. Locking while the request is
in-flight also is not feasible, as this would force _every_ request to
execute in sequence.

Instead of caching the result of the request, we cache its computation.
The first request initiates the request and stores its `Future`, then
all subsequent requests retrieve that `Future` from cache and `await`
its result, ensuring only one underlying request at most.

## Performance Impact

My main concern with this implementation is the impact it will have on
performance. Each request made must block to check the cache,
introducing contention/delays. The lock is only held while checking the
cache, and not while the request is being made, so my hope is that it
does not impact too much. This may be something that needs to be
iterated over time.

From local testing the impact seemed to be negligible, but that was with
5 Indexers, it may be worse with many. I've added a metric to measure
lock wait time, to determine whether this contention is becoming a
problem.