feat: Retry Tracking and Permissions for Tables in Hasura (#663)
Logs Table and Metadata Table are necessary for important functions of
QueryApi. Hasura often invisibly fails to track tables and add
permissions. These operations need to be successful, so I added a check
which verifies tracking and permissions are correct, and reattempts them
if not. When successful, the result is cached. In a successful case, the
expensive hasura calls (getTableNames and
getTrackedTablesWithPermissions) are done at most twice.

I also combined the conditional provisioning functions since they are
verified to work already.