fix: Continue Replacement of Logs/Metadata tables after untracking failures (#675)
Errors in provisioning should be logged to the machine as they can
potentially be overwritten by errors in the finally block of the parent
try catch.

We ideally want to move the provisioning out to its own try catch but
this is a simple fix for the time being.

In addition the PRs to replace the logs/metadata tables failed due to
untracking being partially successful. This PR allows untracking errors.