fix: Remove all hasura write calls to logs/state tables (#692)
Removing writes to old logs and state table now that it is confirmed
that the new tables work as expected. We expect to also see a reduction
in the number of DB connections used by QueryApi since fewer write calls
are being made by Hasura.

I've retained the whitelist so that we can still control provisioning of
new tables.