Implement new Logs Table using Apollo Client (#678)
substituted rest endpoints for logging tables using apollo client. 


https://www.loom.com/share/4dbea64625d2449dbd7b2c6e7e3eecc8?sid=cea1b56c-88a4-488a-8ab1-20cabc76ece4

**Known Issues:**

1. Eslint: React Hook useEffect has a missing dependency: 'renderGrid'.
Either include it or remove the dependency array. This triggers the use
of callbacks, which work fine for fetching/refetching data but cause
conflicts with the Grid.js `search flag` functionality being
responsive..

2. Occasionally, when you click between page, it may trigger a`
'IndexerLogs.jsx:113 [Grid.js] [ERROR]: The container element [object
HTMLDivElement] is not empty. Make sure the container is empty and call
render() again'.` This will not break the page, but numerous attempts to
resolve the issue without direct DOM manipulation have been made. Due to
overriding the form of pagination and adding our own, hidden conflicts
that have not been resolved seem to exist.

I prefer we just move away from gridjs. Outside of its minimal table
formatting we are doing (see gridConfig). The only real features we use
are the search functionality. The library is not inherently bad - Its
extremely lightweight but a poor choice in this case because we use
apollo client as oppose to its more configurable counterpart requiring a
REST endpoint. There is not much control we have when interacting with
gridjs features. As a result - such features and homemade pagination
cause issues due to potential conflicts in managing data rendering and
state updates.

**Odd behavior:**

I only seem to be getting a specific blockheight despite the timestamp
and ID changing so I strongly believe im not fetching the same block.
The IDs are unique but it lead me to think that it seems as we're
repeatedly processing the same block again.

This is what the page looks like if query failed. The console also shows
the error message. The query can mainly fail if either the indexer has
not run once since we first added in provisioning for the new tables
(thus the tables don't exist), or the provisioning failed.
<img width="1582" alt="image"
src="https://github.com/near/queryapi/assets/22734869/65e3d159-f49d-42b9-ba63-dc28618b214e">

---------

Co-authored-by: Darun Seethammagari <darunrs@gmail.com>