fix: Pass startblock regardless of DDL Validation Error (#674)
Passes startblock regardless if DDL Validation Error or not.

The culprit seems to be when we bypass the ddl Validation Error.
` There was an error while generating types for your SQL schema. Please
ensure your schema is valid SQL DDL.`

If the SQL schema is valid we generate the start_block. If its not, we
use a modal with obj and did not pass in the startblock k/v pair. The
modal never had the start_block param.