feat: Introduce wrapSpan utility and Refactor DmlHandler tracing (#649)
Many areas in Runner involve adding a try finally block to support
tracing it. I added a wrapSpan utility function to simplify the code a
little bit. I also refactored DmlHandler's tracing. I was previously
tracing it in Indexer which adds to clutter and isn't the right place
for it. I instead added a private function which does the tracing for
DmlHandler functions. I also add the SQL query as an attribute for the
span.