fix: Catch failed Lake S3 request to prevent unhandled rejections (#725)
Rejected promises must be handled via `await` or `.catch()`, without
either of these Node will exit due to the "unhandled rejection".
Currently, Lake S3 requests are created in advance, and then handled
later when the block is ready to be executed. Usually, this is not an
issue, as the delay between promise creation/handling is not too large.
But for slow indexers (i.e. `nearpavel_near/bitmap_v2`), this delay is
long enough for Node to consider it "unhandled".

This PR attaches a rejection handler to block requests, so that failures
can be handled gracefully. This does not affect the "pre-fetch"
behaviour, these requests will still be executed ahead of time, but
failed requests will be handled within our code. Explicit handling binds
the error to our execution context providing a meaningful call-stack, as
opposed to a lonesome error which seemingly comes from no-where.

To mitigate the problem itself - failed S3 requests, I have bumped
`maxAttempts`, and changed the `retryMode` in the hopes that the
transient error can be overcome.