feat: Store Forked Indexer when Publishing (#724)
Maintaining what Indexer was forked during development is useful for
various usage statistics. I've updated the frontend to now store what
indexer was forked from and store that information in the contract after
publishing. We specifically store the indexer which was directly forked.

With this information, we can construct a graph using the "forked_from"
as edge information, and calculate various other statistics such as how
many indexers used some other indexer as a base by doing path counts.
This can be done by creating an indexer which indexes the QueryApi
contract.

There are some bugs in the workflow currently. Specifically, refreshing
while in the forked page will cause the loss of whatever code was
written. You need to re-fork the Indexer. While this benefits to
ensuring the forked from field is correct, if we end up fixing this
behavior, we need to then ensure forked from is stored persistently.


As part of updating the contract, I've removed various old functions
which are no longer necessary.