refactor: Introduce struct to manager persistent indexer state (#727)
This PR centralises persistent Indexer State within the
`IndexerStateManager` struct. Currently, we only persist the "stream
version", but this will soon grow to "enabled/disabled", which will be
implemented in my next PR. This is just a tidy up to make the next step
a bit easier.

Indexer state will be stored as stringified JSON under
`{account_id}/{function_name}:state`, currently this only includes when
the block stream was last synced. I've included a migration step in to
move from the old key/structure to the new.