fix: Ensure Terminating Thread error is correctly templatized (#738)
The error given to Stream Handler is occasionally a JSON object in
disguise, and not an Error. As a result, calling error.toString()
returns `[Object object]` rather than the error contents. I've added a
check so that if the result of toString() is the earlier value, return a
JSON.stringify result instead. Doing JSON.stringify on a proper Error
type results in `{}`. The two cases must be handled separately.

To test this, I created test indexers which called one of the two pieces
of code which throw unawaited async errors. I verified both wrote the
error message and stack trace into the log table.

```
const timeoutPromise = new Promise((_, reject) => {
  setTimeout(() => {
      reject(new Error('Error thrown after 100ms'));
  }, 100);
});
```
```context.db.IndexerStorage.upsert({}, [], []); ```