feat: Implement Block Streamer Bitmap Operations (#747)
The bitmap indexer will return a list of bitmaps in the form of base 64
strings, and associated start block heights. We need a way to convert
all that data into a single block height and an associated bitmap.

This PR introduces a new BitmapOperator class which holds all the
operations necessary to perform the function of returning a combined
binary bitmap with the lowest start block height as index 0.