feat: Expose gRPC service to handle provisioning (#758)
This PR adds the `DataLayerService`, which allows for provisioning the
"Data Layer" of a QueryAPI Indexer. Data Layer in this case refers to
the Postgres Database and Hasura GraphQL endpoint. I wanted to move away
from the more general "provisioning" term as it is a bit overloaded, and
may get conflated with other provisioning steps related to Executors
(which may become more involved with Firecracker). This PR doesn't
replace the existing provisioning flow - yet, I'll do that in a follow
up PR.

As provisioning can take some time, rather than keeping the
request/connection open while we wait, the `Provision` command triggers
the process, and returns immediately. The `CheckProvsioningStatus`
command will then be used to poll its completion.

On top the new service, the following changes have been made:
- the gRPC `server/` directory has been slightly restructured to
accommodate additional services, i.e. `DataLayerService`, as a result
you'll see lots of `runner` files being moved around
- `Provisioner` now takes `ProvisioningConfig`, which is a subset of
`IndexerConfig`. On the gRPC side we don't have a full `IndexerConfig`,
nor do we need it, so I created the subset to make it possible to
provision with only the data we need. `IndexerConfig extends
ProvisioningConfig` so it's not a breaking change to the existing
use-cases.
- I've removed all the unused code in `Provisioner`