feat: Implement Block Height Stream (#776)
I've implemented a struct which is capable of returning an async stream
which yields block heights matching a particular contract filter
starting from a specified block height. This should serve as a
replacement for the existing method of getting block heights by querying
Delta Lake through the Delta Lake Client.

Some differences between Delta Lake and this new stream is that it
yields a SINGLE block height rather than all block heights. In addition,
we now query the bitmap indexer's tables for block heights for one day,
instead of for all days. This should improve time to get a block as well
as reduce the memory footprint.