fix: Allow re-provisioning of recently de-provisioned resources (#825)
Provisioning a recently de-provisioned Data Layer would silently fail.
This is due to the fact that de/provisioning tasks are stored in-memory,
keyed by a hash of the config. So if a provisioning task was recently
completed, attempting to re-provision would return that same task.

This PR keys by random UUIDs, instead of hashes, so we can trigger
multiple provisioning jobs for a given Indexer/Data Layer, allowing for
the Provision > De-provision > Provision flow. To protect against
re-provisioning an _existing_ Data Layer, we only start the task after
verifying it doesn't already exist. Also removed cache from
`Provisioner` to ensure we are getting accurate results.