feat: Add script to suspend Indexers (#829)
This PR adds a Node script to Runner to suspend Indexers due to
inactivity. The script will:
1. Call Coordinator to disable the indexer
2. Write to the Indexers logs table to notify of suspension

Note that as Coordinator is in a private network, you must tunnel to the
machine to expose the gRPC server. This can be achieved via running the
following in a separate terminal:
```sh
gcloud compute ssh ubuntu@queryapi-coordinator-mainnet -- -L 9003:0.0.0.0:9003
```

The following environment variables are required:
- `HASURA_ADMIN_SECRET`
- `HASURA_ENDPOINT`
- `PGPORT`
- `PGHOST`

All of which can be found in the Runner compute instance metadata:
```sh
gcloud compute instances describe queryapi-runner-mainnet
```


Usage: `npm run script:suspend-indexer -- <accountId> <functionName>`