refactor: Refactor Editor into TypeScript with Smaller Components and Separate Concerns (#830)
Refactored Editor Component to TypeScript. This refactoring involved
breaking down the Editor file into smaller chunks and separating
concerns into distinct components. Also did some minor work on validator
to ts as it is a major consumer in the editor. It is setup to later
iterate on some additional test for validators.