fix: Report worker metrics even when no messages in Stream (#844)
We skip reporting metrics if there are no messages in the pre-fetch
queue/Redis Stream. This is especially problematic for `EXECUTOR_UP`, as
we won't increment the metric even though we are processing.

This PR moves the metrics logic so that it is always reported, even when
no messages in the stream.