fix: Do not eat bitmap backfill errors (#871)
If the stream yielding block heights from the bitmap indexer fails, its
error will be eaten by the while loop as the loop will simply end if any
error takes place. This PR ensures that if the result from the stream is
an error, it logs the error before proceeding with Lake.