fix: fetch BOS indexerdata from near RPC call and conditionally render metadata if available in databricks (#873)
BOS component data is fetched from NEAR RPC call 'list_all'. We use a
map to set a k/v pair for indexer metadata recieved on databricks. When
we render each card we check the map to see if the indexer to be
rendered from NEAR RPC exist in the map. If it does exist we
conditionally render metadata.