feat: In-Memory DmlHandler Test Fixture (#809)
In order to enable quick testing of indexer code for the purposes of
local indexer development, there needs to be a way to functionally mock
dependencies that are used by the Indexer code. These dependencies are
roughly encapsulated under the context object. In particular context.db
is of high importance as it is the method through with Indexers interact
with their persistent data. This PR focuses on creating an MVP of an in
memory DML Handler Test Fixture which can be used interchangeably with
an actual DmlHandler. This allows context.db calls to pretend to be real
calls to an actual Postgres DB without actually having an instance
running (Which would be hard to integrate into unit testing).

This PR is mainly focused on getting a basic DmlHandler test fixture out
the door. It has the same functions which roughly behave similarly to if
they were called on an actual DB. There is the added benefit of being
extremely quick to tear down as all data is represented in memory,
making it suitable for fast iteration through unit testing. However,
since I am essentially mocking Postgres DB responses, the correctness
standard is much lower than using an actual PG DB, but since DmlHandler
simplifies user interactions with the DB anyway, we can mock a great
deal of Indexer context.db use cases sufficiently to serve as useful for
end users.