refactor: Fetch Database Connection Parameters in Stream Handler (#890)
Due to our migration of Provisioning logic to a separate API in Runner,
we can expect that an Indexers resources will be initialized before we
reach the point of executing blocks. So, it is no longer necessary to
lazy init DmlHandler or IndexerMeta due to not having the connection
parameters at the time of class initialization. Worker can poll for
these params before even beginning to execute any blocks. This also
continues to simplify both the code and behavior of the execute
function.