feat: Add metric for receiver block backfill failures (#912)
Adds a metric for tracking receiver block backfill status. Since the
backfill has an end state, I can't rely on constantly incrementing a
counter. Instead, I use a gauge and check for non zero values. If the
backfill starts and completes without issue, the value will be 0. But,
if the backfill fails, it will increment the gauge and then skip to Lake
backfill, leaving the gauge at a nonzero value. We can alert on this. If
the stream is restarted, then a successful attempt at the backfill will
reset the gauge to 0.