feat: Programmatically Execute Indexer Locally (#908)
This PR sets up a local indexer class which leverages the in memory dml
handler class and a no op indexer meta class to fully execute indexer
logic locally on a block. The class is responsible for downloading block
data and providing helper functions to abstract testing logic away. This
PR also sets up a core-indexers folder which is used to verify that the
created local indexer is actually capable of executing indexer code
locally. The core-indexers folder may be temporary as we do intend to
eventually publish a testing framework npm package. For the time being,
it uses relative imports to work.