fix: Implement retry on tracking foreign keys and resolve hasura env issue (#919)
This PR adds exponential retry for the tracking of foreign keys. For now
I'm keeping the implementation of expo retry in provisioner light for
two reasons:
1. I want to confirm that expo retry actually solves the issue.
2. I plan to refactor the provisioning process entirely. 

I also noticed a bug where the env variables were being set to empty in
all cases instead of only if null. I fix this bug as well.