fix: removed overwritting of indexername in obj (#962)
Databricks stores account and function names in their sanitized form
(Only alphanumerics and underscores) as this is how they are stored in
Hasura. However, the actual indexer and accounts are stored in both RPC
and the queryapi_indexer under their original names.

This PR prevents the metadata from databricks overwriting the original
names with the sanitized ones.