fix: Require exact config response for lifecycle transitions (#956)
We've noticed that it is occasionally possible for the indexer config
returned by the contract to be parseable/null with a success status
code, even though the indexer itself was not in fact deleted. This
triggers the deprovisioning workflow anyway, which currently throws the
Indexer into a repairing state, where it is no longer operating.

This PR updates the contract to retain configuration if the indexer was
deleted. Coordinator will now expect a specific configuration to take
deletion actions, instead of the previous method which left things open
to interpretation.

In the near future, the contract will be migrated (#964) to use an enum
instead of contract format, to avoid wastefully storing things like
code/schema for deleted indexers forever.