test: Test fetching indexer config from registry (#976)
Added tests for fetching individual indexer config from registry
contract. This was mostly an exercise to understand the behaviour of
parsing registry contract responses, as well as understanding the
changes that were made recently. A thin wrapper (`JsonRpcClientWrapper`)
has been created to allow mocking of the JSON RPC Client responses.